% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_intervals}
\alias{add_intervals}
\title{Add Intervals to Forecast}
\usage{
add_intervals(forecast, intervals)
}
\arguments{
\item{forecast}{A Forecast object}

\item{intervals}{Interval data structure}
}
\value{
Updated Forecast object
}
\description{
Add Intervals to Forecast
}
