% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_median}
\alias{add_median}
\title{Add Median to Forecast}
\usage{
add_median(forecast, median)
}
\arguments{
\item{forecast}{A Forecast object}

\item{median}{Numeric vector of median forecasts}
}
\value{
Updated Forecast object
}
\description{
Add Median to Forecast
}
