% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_temporal_info}
\alias{add_temporal_info}
\title{Add Temporal Information to Forecast}
\usage{
add_temporal_info(forecast, reference_date, target_date, resolution)
}
\arguments{
\item{forecast}{A Forecast object}

\item{reference_date}{Reference date}

\item{target_date}{Target date}

\item{resolution}{Time resolution}
}
\value{
Updated Forecast object
}
\description{
Add Temporal Information to Forecast
}
