% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_trajectories}
\alias{add_trajectories}
\title{Add Trajectories to Forecast}
\usage{
add_trajectories(forecast, trajectories)
}
\arguments{
\item{forecast}{A Forecast object}

\item{trajectories}{Matrix of trajectories}
}
\value{
Updated Forecast object
}
\description{
Add Trajectories to Forecast
}
