% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_truth}
\alias{add_truth}
\title{Add Truth Values to Forecast}
\usage{
add_truth(forecast, truth)
}
\arguments{
\item{forecast}{A Forecast object}

\item{truth}{Numeric vector of observed values}
}
\value{
Updated Forecast object
}
\description{
Add Truth Values to Forecast
}
