% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{as_forecast_point.ForecastBaselines_Forecast}
\alias{as_forecast_point.ForecastBaselines_Forecast}
\title{Convert ForecastBaselines Forecast to scoringutils point forecast}
\usage{
\method{as_forecast_point}{ForecastBaselines_Forecast}(data, ...)
}
\arguments{
\item{data}{A ForecastBaselines_Forecast object}

\item{...}{Additional arguments (not used)}
}
\value{
A forecast_point object from scoringutils
}
\description{
S3 method to convert a ForecastBaselines_Forecast object to a scoringutils
forecast_point object. This allows seamless integration with scoringutils.
}
\examples{
\dontrun{
# Convert and validate as point forecast
fc_point <- scoringutils::as_forecast_point(forecast)
}
}
