% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ConstantModel}
\alias{ConstantModel}
\title{Constant Model}
\usage{
ConstantModel()
}
\value{
A ConstantModel object
}
\description{
Creates a naive forecast model that uses the last observed value as the forecast
for all future horizons.
}
\examples{
\dontrun{
model <- ConstantModel()
}
}
