% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{CRPS_trajectory}
\alias{CRPS_trajectory}
\title{CRPS from Trajectories}
\usage{
CRPS_trajectory()
}
\value{
A CRPS_trajectory scoring rule object
}
\description{
Creates a CRPS_trajectory scoring rule object. Computes CRPS from
forecast trajectories/samples.
}
\examples{
\dontrun{
rule <- CRPS_trajectory()
score_value <- score(forecast, rule)
}
}
