% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_horizons}
\alias{filter_horizons}
\title{Filter Forecast to Specific Horizons}
\usage{
filter_horizons(forecast, horizons)
}
\arguments{
\item{forecast}{A Forecast object}

\item{horizons}{Vector of horizons to keep}
}
\value{
Filtered Forecast object
}
\description{
Filter Forecast to Specific Horizons
}
