% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_levels}
\alias{filter_levels}
\title{Filter Forecast to Specific Confidence Levels}
\usage{
filter_levels(forecast, levels)
}
\arguments{
\item{forecast}{A Forecast object}

\item{levels}{Vector of confidence levels to keep}
}
\value{
Filtered Forecast object
}
\description{
Filter Forecast to Specific Confidence Levels
}
