% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{fit_baseline}
\alias{fit_baseline}
\title{Fit a Baseline Model}
\usage{
fit_baseline(x, model, temporal_info = NULL)
}
\arguments{
\item{x}{Numeric vector of time series data}

\item{model}{A model object created by one of the model constructors}

\item{temporal_info}{Optional TemporalInfo object with start date and resolution}
}
\value{
A fitted model object that can be used for forecasting
}
\description{
Fits a baseline forecasting model to observed data.
}
\examples{
\dontrun{
data <- c(1.2, 2.3, 3.1, 2.8, 3.5, 4.2, 3.9)
model <- ARMAModel(p = 1, q = 1)
fitted <- fit_baseline(data, model)
}
}
