% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_horizon}
\alias{has_horizon}
\title{Check if Forecast has Horizon}
\usage{
has_horizon(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Horizon
}
