% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_intervals}
\alias{has_intervals}
\title{Check if Forecast has Intervals}
\usage{
has_intervals(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Intervals
}
