% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_mean}
\alias{has_mean}
\title{Check if Forecast has Mean}
\usage{
has_mean(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Mean
}
