% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_median}
\alias{has_median}
\title{Check if Forecast has Median}
\usage{
has_median(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Median
}
