% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_trajectories}
\alias{has_trajectories}
\title{Check if Forecast has Trajectories}
\usage{
has_trajectories(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Trajectories
}
