% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_truth}
\alias{has_truth}
\title{Check if Forecast has Truth Values}
\usage{
has_truth(forecast)
}
\arguments{
\item{forecast}{A Forecast object}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if Forecast has Truth Values
}
