% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{INARCHModel}
\alias{INARCHModel}
\title{INARCH Model}
\usage{
INARCHModel(p = 1L)
}
\arguments{
\item{p}{Order of the INARCH model}
}
\value{
An INARCHModel object
}
\description{
Creates an Integer-valued ARCH model for count time series.
}
\examples{
\dontrun{
model <- INARCHModel(p = 1)
}
}
