% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{is_setup}
\alias{is_setup}
\title{Check if Julia and ForecastBaselines.jl are set up}
\usage{
is_setup()
}
\value{
TRUE if setup is complete, FALSE otherwise
}
\description{
Check if Julia and ForecastBaselines.jl are set up
}
