% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{LogTransform}
\alias{LogTransform}
\title{Log Transformation}
\usage{
LogTransform()
}
\value{
A LogTransform object
}
\description{
Creates a natural logarithm transformation. Data must be positive.
}
\examples{
\dontrun{
trans <- LogTransform()
}
}
