% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{MAPE}
\alias{MAPE}
\title{Mean Absolute Percentage Error (MAPE)}
\usage{
MAPE()
}
\value{
A MAPE scoring rule object
}
\description{
Creates a MAPE scoring rule object. MAPE measures the average absolute
percentage error.
}
\examples{
\dontrun{
rule <- MAPE()
score_value <- score(forecast, rule)
}
}
