% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{MarginalModel}
\alias{MarginalModel}
\title{Marginal Model}
\usage{
MarginalModel(p = NULL)
}
\arguments{
\item{p}{Number of most recent observations to use (default: all observations)}
}
\value{
A MarginalModel object
}
\description{
Creates a forecast based on the empirical marginal distribution using the mean
of the p most recent observations.
}
\examples{
\dontrun{
model <- MarginalModel(p = 10)
}
}
