% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{MdAE}
\alias{MdAE}
\title{Median Absolute Error (MdAE)}
\usage{
MdAE()
}
\value{
An MdAE scoring rule object
}
\description{
Creates an MdAE scoring rule object. MdAE measures the median absolute
difference between forecasts and observations.
}
\examples{
\dontrun{
rule <- MdAE()
score_value <- score(forecast, rule)
}
}
