% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{MSPE}
\alias{MSPE}
\title{Mean Squared Percentage Error (MSPE)}
\usage{
MSPE()
}
\value{
An MSPE scoring rule object
}
\description{
Creates an MSPE scoring rule object. MSPE measures the average squared
percentage error.
}
\examples{
\dontrun{
rule <- MSPE()
score_value <- score(forecast, rule)
}
}
