% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{NoInterval}
\alias{NoInterval}
\title{No Interval Method}
\usage{
NoInterval()
}
\value{
A NoInterval object
}
\description{
Creates an interval method that produces only point forecasts without
prediction intervals.
}
\examples{
\dontrun{
method <- NoInterval()
fc <- forecast(fitted, interval_method = method, horizon = 1:12)
}
}
