% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{NoTransform}
\alias{NoTransform}
\title{No Transformation}
\usage{
NoTransform()
}
\value{
A NoTransform object
}
\description{
Creates a transformation that applies no changes to the data.
}
\examples{
\dontrun{
trans <- NoTransform()
}
}
