% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{PowerTransform}
\alias{PowerTransform}
\title{Power Transformation}
\usage{
PowerTransform(lambda)
}
\arguments{
\item{lambda}{Power parameter (lambda = 0 is log, lambda = 0.5 is sqrt)}
}
\value{
A PowerTransform object
}
\description{
Creates a power transformation: x^lambda (Box-Cox family).
}
\examples{
\dontrun{
# Square root (equivalent to SquareRootTransform)
trans <- PowerTransform(lambda = 0.5)

# Cube root
trans <- PowerTransform(lambda = 1 / 3)
}
}
