% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{print.ForecastBaselines_Forecast}
\alias{print.ForecastBaselines_Forecast}
\title{Print method for Forecast objects}
\usage{
\method{print}{ForecastBaselines_Forecast}(x, ...)
}
\arguments{
\item{x}{A Forecast object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for Forecast objects
}
