% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{TemporalInfo}
\alias{TemporalInfo}
\title{Create Temporal Information Object}
\usage{
TemporalInfo(start = 1, resolution = 1)
}
\arguments{
\item{start}{Start date (Date, POSIXct, or integer)}

\item{resolution}{Time resolution (integer for generic, or period object)}
}
\value{
A TemporalInfo object
}
\description{
Creates a TemporalInfo object to track dates and time resolution.
}
\examples{
\dontrun{
# Simple integer indexing
temp_info <- TemporalInfo(start = 1, resolution = 1)

# Date-based
temp_info <- TemporalInfo(start = as.Date("2024-01-01"), resolution = 1)
}
}
