% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{transform_model}
\alias{transform_model}
\title{Apply Transformation to Model}
\usage{
transform_model(model, transformation)
}
\arguments{
\item{model}{A model object}

\item{transformation}{A transformation object}
}
\value{
A transformed model object
}
\description{
Wraps a model with a data transformation.
}
\details{
**Not Implemented:** This function does not work because ForecastBaselines.jl
does not implement `transform()` for model types.

**Recommended approach:** Transform your data manually in R before fitting:

```r
# Instead of transform_model():
log_data <- log(data)
fitted <- fit_baseline(log_data, model)
fc <- forecast(fitted, ...)
fc$mean <- exp(fc$mean)  # Back-transform
```

See `vignette("transformations")` for complete examples.
}
\examples{
\dontrun{
# This will fail - not implemented in Julia package
model <- ARMAModel(p = 1, q = 1)
trans <- LogTransform()
transformed_model <- transform_model(model, trans) # Error!

# Use manual transformation instead:
log_data <- log(data)
fitted <- fit_baseline(log_data, model)
fc <- forecast(fitted, interval_method = NoInterval(), horizon = 1:5)
fc$mean <- exp(fc$mean)
}
}
