% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_horizon}
\alias{truncate_horizon}
\title{Truncate Forecast Horizon}
\usage{
truncate_horizon(forecast, max_h)
}
\arguments{
\item{forecast}{A Forecast object}

\item{max_h}{Maximum horizon to keep}
}
\value{
Truncated Forecast object
}
\description{
Truncate Forecast Horizon
}
