% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{is_setup}
\alias{is_setup}
\title{Check if Julia and ForecastBaselines.jl are set up}
\usage{
is_setup()
}
\value{
TRUE if setup is complete, FALSE otherwise
}
\description{
Tests whether Julia is configured and ForecastBaselines.jl is accessible.
}
\examples{
\dontrun{
if (is_setup()) {
  # Use forecasting functions
} else {
  setup_ForecastBaselines()
}
}
}
