% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoring.R
\name{RMSE}
\alias{RMSE}
\title{Root Mean Squared Error (RMSE)}
\usage{
RMSE()
}
\value{
An RMSE scoring rule object
}
\description{
Creates an RMSE scoring rule object. RMSE is the square root of MSE.
}
\examples{
\dontrun{
rule <- RMSE()
score_value <- score(forecast, rule)
}
}
